Precision mixta simetrica en ambos frameworks, apagada por defecto
PyTorch usa autocast + GradScaler; Keras usa la policy mixed_float16.
Son equivalentes: variables en float32, computo en float16, con escalado
de la loss para que los gradientes chicos no se anulen. La capa de salida
se fuerza a float32 por estabilidad numerica del softmax.

Queda apagada por defecto porque solo acelera si la GPU es el cuello de
botella: si esta esperando datos no cambia nada y agrega riesgo numerico.
preflight.sh ahora mide con y sin para decidir con datos, y run_all.sh la
expone via USE_AMP=1.

Verificado en ambos frameworks que sin GPU avisa y cae a float32 sin
romper, y que las metricas registran si se uso.

+++ utils/model_defs_tensorflow.py
@@ -79,7 +79,9 @@ def build_simple_cnn(
         layers.Dense(256, kernel_initializer=init),
         layers.ReLU(),
         layers.Dropout(dropout),
-        layers.Dense(num_classes, kernel_initializer=init),  # logits, igual que la version PyTorch
+        # dtype float32 explicito: con la policy mixed_float16 la salida debe quedar en
+        # float32 por estabilidad numerica del softmax y la loss. Sin la policy es un no-op.
+        layers.Dense(num_classes, kernel_initializer=init, dtype="float32"),  # logits
     ]
     return tf.keras.Sequential(steps)
